feat: 구독 취소 이벤트 기반으로 동작하도록 수정

+++ backend/bom-bom-server/src/main/java/me/bombom/api/v1/subscribe/service/SubscribeService.java
@@ -9,9 +9,12 @@ import me.bombom.api.v1.common.exception.ErrorDetail;
 import me.bombom.api.v1.common.exception.UnauthorizedException;
 import me.bombom.api.v1.member.domain.Member;
 import me.bombom.api.v1.subscribe.domain.Subscribe;
-import me.bombom.api.v1.subscribe.dto.UnsubscribeResponse;
+import me.bombom.api.v1.subscribe.domain.SubscribeStatus;
 import me.bombom.api.v1.subscribe.dto.SubscribedNewsletterResponse;
+import me.bombom.api.v1.subscribe.dto.UnsubscribeResponse;
+import me.bombom.api.v1.subscribe.event.UnsubscribeRequestedEvent;
 import me.bombom.api.v1.subscribe.repository.SubscribeRepository;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,6 +26,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class SubscribeService {
 
     private final SubscribeRepository subscribeRepository;
+    private final ApplicationEventPublisher applicationEventPublisher;
 
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     public void deleteAllByMemberId(Long memberId) {
@@ -49,9 +53,23 @@ public class SubscribeService {
                 .addContext("subscribeId", subscribeId);
         }
 
-        String unsubscribeUrl = subscribe.getUnsubscribeUrl();
-        subscribeRepository.delete(subscribe);
+        // 사용자가 직접 구독 취소 후 삭제 버튼 클릭
+        if (subscribe.isFailed()) {
+            log.info("구독 취소 실패 상태인 항목 강제 삭제 subscribeId: {}", subscribeId);
+            subscribeRepository.delete(subscribe);
+            return UnsubscribeResponse.of(subscribe.getUnsubscribeUrl());
+        }
 
-        return UnsubscribeResponse.of(unsubscribeUrl);
+        // 구독 취소 이미 진행 중
+        if (subscribe.isUnsubscribing()) {
+            return UnsubscribeResponse.of(subscribe.getUnsubscribeUrl());
+        }
+
+        subscribe.changeStatus(SubscribeStatus.UNSUBSCRIBING);
+        applicationEventPublisher.publishEvent(new UnsubscribeRequestedEvent(
+                subscribe.getId(),
+                subscribe.getUnsubscribeUrl()
+        ));
+        return UnsubscribeResponse.of(subscribe.getUnsubscribeUrl());
     }
 }
